fix: correct datetime encoding to match 2024 NSIDC-accepted spec (#4)
NSIDC rejected the 2025 delivery due to incorrect datetime variable encoding.
Root cause: scene_1_datetime and scene_2_datetime were written as float64
instead of int64, which does not conform to the CF convention for time
coordinates stored as integer seconds.

Fix applied to both 4b (Landsat) and 4c (combined Landsat+Sentinel-2):
  - scene_1_datetime: float64 → int64, no _FillValue (matches 2024)
  - scene_2_datetime: float64 → int64, no _FillValue (matches 2024)
  - midpoint_datetime: kept as float64 with _FillValue=NaN (matches 2024)

Note: midpoint_datetime float64 produces a UserWarning (sub-second precision)
at write time. This is expected — the 2024 NSIDC-accepted delivery had the
same warning and was accepted. Switching midpoint to int64 is a potential
future enhancement but requires team discussion with NSIDC before adoption.

+++ 3_orthocorrect_and_netcdf-package/processing_chain/4b_netcdf_stack_landsat.py
@@ -478,11 +478,15 @@ log_to_stdout_and_file(f"Exporting... (Landsat-8 memory size: {ds.nbytes / (1024
 # Get encoding information.
 comp = dict(zlib=True, complevel=1)
 encoding_comp = {var: comp for var in ds}
-# Feb 19, 2026 (BNY): appending , "dtype": "float64" to prevent user warning (see issue 3)
+# Apr 10, 2026 (BNY): scene_1/2_datetime reverted to int64 — this is what 2024 NSIDC-accepted files use.
+# midpoint_datetime intentionally kept as float64 to match the 2024 reference exactly.
+# Note: float64 midpoint may trigger a UserWarning (sub-second precision) — this is acceptable;
+# the 2024 delivery had the same and was accepted. Changing to int64 is a potential future enhancement
+# but requires team discussion with NSIDC before adoption. (see troubleshooting.md issue #2, #2b)
 encoding_time = {
-    "scene_1_datetime": {"units": "seconds since 1970-01-01 00:00:00", "dtype": "float64"},
-    "scene_2_datetime": {"units": "seconds since 1970-01-01 00:00:00", "dtype": "float64"},
-    "midpoint_datetime": {"units": "seconds since 1970-01-01 00:00:00", "dtype": "float64"},
+    "scene_1_datetime": {"units": "seconds since 1970-01-01 00:00:00", "dtype": "int64"},
+    "scene_2_datetime": {"units": "seconds since 1970-01-01 00:00:00", "dtype": "int64"},
+    "midpoint_datetime": {"units": "seconds since 1970-01-01 00:00:00", "dtype": "float64"},  # matches 2024; see note above
 }
 encoding = {**encoding_comp, **encoding_time}
